Updated model dtype from fp32 to bf16

+++ mammoth/hf_integration/from_hf/gemma3/convert_gemma3_decoder_only.py
@@ -81,7 +81,7 @@ def build_model_opts(config):
 
     opts = Namespace()
     opts.seed = 1
-    opts.model_dtype = "fp32"
+    opts.model_dtype = "bf16"
     opts.log_model_structure = False
     opts.decoder_only = True
     # Always weighted_sampling (see build_task_queue_manager) -- recorded
@@ -284,7 +284,7 @@ def convert(
 
     config = get_text_config(hf_model_path)
     hf_model = AutoModelForCausalLM.from_pretrained(
-        hf_model_path, torch_dtype=torch.float32, local_files_only=True,
+        hf_model_path, dtype=torch.bfloat16, local_files_only=True,
     ).eval()
 
     tokenizer_path = prepare_text_only_tokenizer(hf_model_path, config.vocab_size)
